% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auLinTest.R
\name{autoLinTest}
\alias{autoLinTest}
\title{autoLinTest}
\usage{
autoLinTest(
  data,
  y,
  train_subset = c(rep(1, nrow(data))),
  exclude = NULL,
  include = NULL,
  bonus_orig = 0.05,
  show_r2 = 0.7,
  method_cor = "pearson",
  n_qts = 10,
  pct_na_allowed = 0.9,
  quantil_variability_exclude = 0.015,
  trim = 0.01,
  warn = NULL,
  subsample_rows = 1,
  seed = NULL,
  debug = FALSE
)
}
\arguments{
\item{data}{R data frame object.}

\item{y}{Name of you target column.}

\item{train_subset}{Vector of 0 and 1, with the same len of data.}

\item{exclude}{If you do not want to pass the name of the Data Frame variables to be analyzed by the algorithm, you also have to pass the list of data frame variables that are not analyzed, for example the key variables, and everything else will be analyzed.}

\item{include}{List with name of variables to be analyzed by the algorithm.}

\item{bonus_orig}{Bonus give to original var.}

\item{show_r2}{Min value of R2 to print.}

\item{method_cor}{Method to calculate the correlation with the y.}

\item{n_qts}{Number of points the function will try to cut your var to analyze.}

\item{pct_na_allowed}{Pct of NA allowed in each covariable.}

\item{quantil_variability_exclude}{Pct of qunatil, to remove vars that not present variability before discretize.}

\item{trim}{Pct of trim vars, to remove outliers before discretize.}

\item{warn}{Indicate if function will show warnings.}

\item{subsample_rows}{Proportion of your data that function will run.}

\item{seed}{Seed to subsample rows.}

\item{debug}{debug.}
}
\value{
Dont return anything, just print the R squares.
}
\description{
This function auto transfrom  your vars and test the linearity for regression models.
}
\examples{
\dontrun{
autoLinTest(data = data
           ,y = "FPD"
           ,train_subset = data$Flag_DES
           ,include = include
           ,show_cor = 0.7
)
}
}
